First Swift App
My first time trying to design a SWIFT app. The purpose of the app was to connect it with the 4ME API, to use 4ME without actually using the 4ME website. Later on we found out about 4me self service and decided to cancel the project. The guiApp project is 75% done. I focused mostly on the front end and wanted to work with the back end after finishing it. The webApp project is just a "browser" that opens the self-service website.